Struct xfs::BlockMapBTree

pub struct BlockMapBTree {
    pub lookups: u32,
    pub compares: u32,
    pub inserts: u32,
    pub deletes: u32,
}

Fields

lookups

Number of block map btree lookup operations on XFS files.

compares

Number of block map btree compare operations in XFS block map lookups.

inserts

Number of block map btree records inserted for XFS files.

deletes

Number of block map btree records deleted for XFS files.