Struct xfs::BlockMapping

pub struct BlockMapping {
    pub map_read: u32,
    pub map_write: u32,
    pub unmap: u32,
    pub list_insert: u32,
    pub list_delete: u32,
    pub list_lookup: u32,
    pub list_compare: u32,
}

Fields

map_read

Number of block map for read operations performed on XFS files.

map_write

Number of block map for write operations performed on XFS files.

unmap

Number of block unmap (delete) operations performed on XFS files.

list_insert

Number of extent list insertion operations for XFS files.

list_delete

Number of extent list deletion operations for XFS files.

list_lookup

Number of extent list lookup operations for XFS files.

list_compare

Number of extent list comparisons in XFS extent list lookups.