Struct xfs::ExtendedPrecisionCounters

pub struct ExtendedPrecisionCounters {
    pub xstrat_bytes: u64,
    pub write_bytes: u64,
    pub read_bytes: u64,
}

Fields

xstrat_bytes

This is a count of bytes of file data flushed out by the XFS flushing daemons.

write_bytes

This is a count of bytes written via write(2) system calls to files in XFS file systems. It can be used in conjunction with the write_calls count to calculate the average size of the write operations to files in XFS file systems.

read_bytes

This is a count of bytes read via read(2) system calls to files in XFS file systems. It can be used in conjunction with the read_calls count to calculate the average size of the read operations to files in XFS file systems.