Struct xfs::TailPushingStats

pub struct TailPushingStats {
    pub logspace: u32,
    pub sleep_logspace: u32,
    pub push_ails: u32,
    pub push_ail_success: u32,
    pub push_ail_pushbuf: u32,
    pub push_ail_pinned: u32,
    pub push_ail_locked: u32,
    pub push_ail_flushing: u32,
    pub push_ail_restarts: u32,
    pub push_ail_flush: u32,
}

Fields

logspace

Value from the xs_try_logspace field of struct xfsstats.

sleep_logspace

Value from the xs_sleep_logspace field of struct xfsstats.

push_ails

The number of times the tail of the AIL is moved forward. It is equivalent to the number of successful calls to the function xfs_trans_push_ail().

push_ail_success

Value from xs_push_ail_success field of struct xfsstats.

push_ail_pushbuf

Value from xs_push_ail_pushbuf field of struct xfsstats.

push_ail_pinned

Value from xs_push_ail_pinned field of struct xfsstats.

push_ail_locked

Value from xs_push_ail_locked field of struct xfsstats.

push_ail_flushing

Value from xs_push_ail_flushing field of struct xfsstats.

push_ail_restarts

Value from xs_push_ail_restarts field of struct xfsstats.

push_ail_flush

Value from xs_push_ail_flush field of struct xfsstats.